Corrected the location parameter of all the schemas.
What was the problem:
The location parameter for all the schemas was defined simply by having the latitude and longitde. This was not correct as the location parameter is a GeoJSON object and should be defined as such. This commit corrects that by implementing the proper GeoJSON object for the location parameter.
Changes made:
- Changed the location parameter of the schemas to a GeoJSON object.
- Made the location parameter required for all the schemas.
- The location parameter for all the schemas is of 'Point' type.
- The coordinates of the location parameter are an array of two Decimal128 values, storing the LONGITDE and LATITUDE respectively.