Added a locationLog schema to the DB_Models folder.
Created and defined a new schema for the location logs of the user devices to be stored in the database. This database will be used to track the location of the user devices and location data will be send to the devices to be displayed on the map. This will also be a part of the analytical data stored for the Application.