Added the locationLog model controller and routes.
The controllers and routes for handling location log requests were added in this commit. The references for these were added to the main server and the database model index file.
Proceeding forward, the locations of users will be processed in database. Also the log database references the user database to retrieve the user details, if needed.